+++ src/main/java/fuzs/visualworkbench/world/level/block/CraftingBlockEntityProvider.java
@@ -1,5 +1,0 @@
-package fuzs.visualworkbench.world.level.block;
-
-public interface CraftingBlockEntityProvider {
-    boolean withCraftingBlockEntity();
-}