+++ CHANGELOG.md
@@ -3,13 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v2.0.0-1.17.1] - 2021-12-06
+### Added
+- Added new system for manually registering crafting table blocks that should be supported by Visual Workbench
+- Previous solution was too unreliable
+### Fixed
+- Fixed a client desync issue when FastWorkbench Minus Replacement mod is installed
+
 ## [v2.0.0b2-1.17.1] - 2021-11-28
 ### Fixed
 - Fixed an issue where shift-clicking multiple crafting results wouldn't work properly
 
 ## [v2.0.0b1-1.17.1] - 2021-11-23
 - Ported to Minecraft 1.17
-### Issues
-- Fletching and smithing tables cannot be moved by pistons
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/